Use separate namespaces for check categories
