+++ tests/phpunit/tests/Checker/Checks/Enqueued_Scripts_Size_Check_Tests.php
@@ -5,7 +5,7 @@
  * @package plugin-check
  */
 
-use WordPress\Plugin_Check\Checker\Checks\Enqueued_Scripts_Size_Check;
+use WordPress\Plugin_Check\Checker\Checks\Performance\Enqueued_Scripts_Size_Check;
 use WordPress\Plugin_Check\Checker\Preparation;
 use WordPress\Plugin_Check\Test_Utils\TestCase\Runtime_Check_UnitTestCase;
