+++ tests/phpunit/tests/Checker/Checks/Late_Escaping_Check_Tests.php
@@ -7,7 +7,7 @@
 
 use WordPress\Plugin_Check\Checker\Check_Context;
 use WordPress\Plugin_Check\Checker\Check_Result;
-use WordPress\Plugin_Check\Checker\Checks\Late_Escaping_Check;
+use WordPress\Plugin_Check\Checker\Checks\Security\Late_Escaping_Check;
 
 class Late_Escaping_Check_Tests extends WP_UnitTestCase {
